% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{isStringBlank}
\alias{isStringBlank}
\title{input is null or empty string}
\usage{
isStringBlank(input)
}
\arguments{
\item{input}{}
}
\value{
bool
}
\description{
input is null or empty string
}
\examples{
isStringBlank(aaa)
}
