% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-utils.R
\name{isValueEmpty}
\alias{isValueEmpty}
\title{input is null, na, or empty string}
\usage{
isValueEmpty(input)
}
\arguments{
\item{input}{}
}
\value{
bool
}
\description{
input is null, na, or empty string
}
\examples{
isValueEmpty(x)
}
