% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-utils.R
\name{readDataset}
\alias{readDataset}
\title{通过文件名读取数据}
\usage{
readDataset(dataset_path)
}
\arguments{
\item{dataset_path}{}
}
\value{
dataset
}
\description{
通过文件名读取数据
}
\examples{
dataset <- readDataset("/path/ggplot2_mpg.rds")
}
